+++ handlers_test.go
@@ -1,0 +1,40 @@
+package main
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
+)
+
+func TestHealthCheckHandler(t *testing.T) {
+	// Create a request to pass to the handler.
+	req, err := http.NewRequest("GET", "/healthz", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a ResponseRecorder to capture the response from the handler.
+	rr := httptest.NewRecorder()
+
+	// Call the health check handler, passing in the request and ResponseRecorder.
+	healthCheckHandler(rr, req)
+
+	// Check the HTTP status code.  We expect 200 OK.
+	status := rr.Code
+	if status != http.StatusOK {
+		t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusOK)
+	}
+
+	// Check the Content-Type header. We expect "application/json".
+	contentType := rr.Header().Get("Content-Type")
+	if contentType != "application/json" {
+		t.Errorf("handler returned wrong content type: got %v want %v", contentType, "application/json")
+	}
+
+	// Check the response body. We expect `{"Health":"OK"}`.
+	expected := `{"Health":"OK"}`
+	actual := rr.Body.String()
+	if actual != expected {
+		t.Errorf("handler returned unexpected body: got %v want %v", actual, expected)
+	}
+}
